+++ git-ref/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Replace `transaction::Create` with `transaction::PreviousValue` and remove `transaction::Create`
 * Remove `file::Reference` in favor of `Reference`
 * Move `file::log::Line` to `log::Line`
+* `TargetRef::Symbolic(&BStr)` -> `TargetRef::Symbolic(FullNameRef)`
 
 ### v0.6.1